+++ source/DSES/ExecutionControl.cpp
@@ -425,8 +425,8 @@ void ExecutionControl::wait_for_all_multiphase_init_sync_points()
            && ( sp->label.compare( DSES::INITIALIZE_SYNC_POINT ) != 0 )
            && ( sp->label.compare( DSES::SIM_CONFIG_SYNC_POINT ) != 0 ) ) {
 
-         long long    wallclock_time;
-         SleepTimeout print_timer( (double)federate->wait_status_time );
+         int64_t      wallclock_time;
+         SleepTimeout print_timer( federate->wait_status_time );
          SleepTimeout sleep_timer;
 
          // Wait for the federation to synchronized on the sync-point.
