+++ source/TrickHLA/ExecutionConfigurationBase.cpp
@@ -249,8 +249,8 @@ void ExecutionConfigurationBase::wait_for_registration()
    bool any_unregistered_obj;
    int  total_obj_cnt = 1;
 
-   long long    wallclock_time;
-   SleepTimeout print_timer( (double)federate->wait_status_time );
+   int64_t      wallclock_time;
+   SleepTimeout print_timer( federate->wait_status_time );
    SleepTimeout sleep_timer;
 
    do {
@@ -359,8 +359,8 @@ bool ExecutionConfigurationBase::wait_for_update() // RETURN: -- None.
    // Make sure we have at least one piece of Execution Configuration data we can receive.
    if ( this->any_remotely_owned_subscribed_init_attribute() ) {
 
-      long long    wallclock_time;
-      SleepTimeout print_timer( (double)federate->wait_status_time );
+      int64_t      wallclock_time;
+      SleepTimeout print_timer( federate->wait_status_time );
       SleepTimeout sleep_timer;
 
       // Wait for the data to arrive.
