+++ source/TrickHLA/SyncPntListBase.cpp
@@ -288,8 +288,8 @@ void SyncPntListBase::wait_for_list_synchronization(
    Federate *federate )
 {
    bool         achieved, valid_and_not_achieved;
-   long long    wallclock_time;
-   SleepTimeout print_timer( (double)federate->wait_status_time );
+   int64_t      wallclock_time;
+   SleepTimeout print_timer( federate->wait_status_time );
    SleepTimeout sleep_timer;
 
    // Iterate through this SyncPntList's synchronization point list.
@@ -806,8 +806,8 @@ bool SyncPntListBase::wait_for_sync_point_announcement(
       }
 
       bool         print_summary = DebugHandler::show( DEBUG_LEVEL_4_TRACE, DEBUG_SOURCE_FEDERATE );
-      long long    wallclock_time;
-      SleepTimeout print_timer( (double)federate->wait_status_time );
+      int64_t      wallclock_time;
+      SleepTimeout print_timer( federate->wait_status_time );
       SleepTimeout sleep_timer;
 
       // Wait for the sync-point to be announced.
@@ -948,8 +948,8 @@ bool SyncPntListBase::wait_for_synchronization(
    if ( sp != NULL ) {
       bool         print_summary = DebugHandler::show( DEBUG_LEVEL_2_TRACE, DEBUG_SOURCE_FEDERATE );
       bool         synchronized;
-      long long    wallclock_time;
-      SleepTimeout print_timer( (double)federate->wait_status_time );
+      int64_t      wallclock_time;
+      SleepTimeout print_timer( federate->wait_status_time );
       SleepTimeout sleep_timer;
 
       // Wait for the federation to synchronize on the sync-point.
